Start BeginMainFrame on empty documents
This change removes a dubious line in
LocalFrameView::BeginLifecycleUpdates() where we never
began updates for empty documents. There is no reason for this,
as empty documents would not cost anything to update and for the
case of certain content (empty iframes, some window.open() cases)
it results in script never executing to populate the frame.

Fixed: 1231231
Change-Id: I51af46888f844a9f5425d03c14e69ef8cacf924e