Remove VRDisplayInfo and pass initial views to XRSessionDeviceConfig
VRDisplayInfo currently only has one field - the initial views when the
session starts. This field is now moved into XRSessionDeviceConfig and
VRDisplayInfo is removed completely. These initial views are only used
before any frames are requested. Once frames are requested, updated
views are sent with each XRFrame.

The initial views are still needed because various places assume that
views exist before frames are requested, for example when initializing
a XRWebGLLayer or if the browser renderer needs to take over due to
frames not being requested by the web page. This can be refactored in
a future change if appropriate.

Bug: 862803
Change-Id: Ibba16ff71c0b8da0a39e5a55fcfcc397246ea36e