Check for 'revert-layer' in both versions of IsCSSWideKeyword.
This fixes the inconsistency between two overloads of IsCSSWideKeyword,
only one of which checked for 'revert-layer'.

The added test in
css/css-properties-values-api/register-property-syntax-parsing.html
fails without the change (OR without the existing codepath; it seems to
require both).

The added test that tests "medium revert-layer" in
css/css-font-loading/fontfaceset-load-css-wide-keywords.html is also
fixed by this change.  (The added test that tests only "revert-layer" is
unaffected; that's why I added the set testing with "medium " prefixed.)

Fixed: 1340993
Change-Id: I12e36c5b8db1e51647e1fcbfb904c8db8fe27a59